Move max time step computation
