+++ polaris/ocean/tasks/barotropic_gyre/forward.py
@@ -3,7 +3,6 @@ from math import ceil
 
 from polaris.mesh.planar import compute_planar_hex_nx_ny
 from polaris.ocean.model import OceanModelStep, get_time_interval_string
-from polaris.ocean.tasks.barotropic_gyre.resources import compute_max_time_step
 
 
 class Forward(OceanModelStep):
@@ -154,3 +153,14 @@ class Forward(OceanModelStep):
         # make sure output is double precision
         self.add_yaml_file(self.package, self.yaml_filename,
                            template_replacements=replacements)
+
+
+def compute_max_time_step(config):
+    u_max = 1  # m/s
+    stability_parameter_max = 0.25
+    resolution = config.getfloat('barotropic_gyre', 'resolution')
+    f_0 = config.getfloat("barotropic_gyre", "f_0")
+    dt_max = min(stability_parameter_max * resolution * 1e3 /
+                 (2 * u_max),
+                 stability_parameter_max / f_0)
+    return dt_max
